backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
psycopg2-binary==2.9.9
pyjwt==2.8.0
bcrypt==4.1.3
python-dotenv==1.0.1
requests==2.31.0
